페이징 완료

+++ src/main/java/com/example/giup1/dto/ProductResponseDto.java
@@ -8,10 +8,10 @@ import java.util.List;
 public class ProductResponseDto {
     private final int totalCount;
     private final Float score;
-    private final int cursor;
+    private final Long cursor;
     private final List<ResponseDto> reviews;
 
-    public ProductResponseDto(int totalCount, Float score, int cursor, List<ResponseDto> reviews) {
+    public ProductResponseDto(int totalCount, Float score, Long cursor, List<ResponseDto> reviews) {
         this.totalCount = totalCount;
         this.score = score;
         this.cursor = cursor;
